+++ src/encode.c
@@ -578,7 +578,7 @@ PyImaging_XbmEncoderNew(PyObject* self, PyObject* args)
 
 #ifdef HAVE_LIBZ
 
-#include "ZipState.h"
+#include "ZipCodecs.h"
 
 PyObject*
 PyImaging_ZipEncoderNew(PyObject* self, PyObject* args)
